# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for a full-stack calculator. Exposes basic and advanced",
        "title": "Sezzle Calculator API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/calculate/add": {
            "post": {
                "description": "Returns a + b",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Add two numbers",
                "parameters": [
                    {
                        "description": "Operands",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_santigonzalezla_sezzle-calculator_internal_httpserver.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/calculate/divide": {
            "post": {
                "description": "Returns a / b. Fails with 400 if b is 0.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Divide two numbers",
                "parameters": [
                    {
                        "description": "Operands",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateResponse"
                        }
                    },
                    "400": {
                        "description": "b must not be zero",
                        "schema": {
                            "$ref": "#/definitions/github_com_santigonzalezla_sezzle-calculator_internal_httpserver.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/calculate/multiply": {
            "post": {
                "description": "Returns a * b",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Multiply two numbers",
                "parameters": [
                    {
                        "description": "Operands",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_santigonzalezla_sezzle-calculator_internal_httpserver.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/calculate/percentage": {
            "post": {
                "description": "Returns a percent of b (e.g. a=50, b=200 -\u003e 100)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Percentage",
                "parameters": [
                    {
                        "description": "Operands",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_santigonzalezla_sezzle-calculator_internal_httpserver.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/calculate/power": {
            "post": {
                "description": "Returns a^b. Fails with 400 if the result is not finite.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Raise a to the power of b",
                "parameters": [
                    {
                        "description": "Base (a) and exponent (b)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_santigonzalezla_sezzle-calculator_internal_httpserver.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/calculate/sqrt": {
            "post": {
                "description": "Returns sqrt(a). B is ignored. Fails with 400 if a is negative.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Square root of a number",
                "parameters": [
                    {
                        "description": "Only a is used",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateResponse"
                        }
                    },
                    "400": {
                        "description": "a must not be negative",
                        "schema": {
                            "$ref": "#/definitions/github_com_santigonzalezla_sezzle-calculator_internal_httpserver.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/calculate/subtract": {
            "post": {
                "description": "Returns a - b",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Subtract two numbers",
                "parameters": [
                    {
                        "description": "Operands",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_calculator.CalculateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_santigonzalezla_sezzle-calculator_internal_httpserver.ErrorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_santigonzalezla_sezzle-calculator_internal_httpserver.ErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_santigonzalezla_sezzle-calculator_internal_httpserver.ErrorDetail"
                }
            }
        },
        "github_com_santigonzalezla_sezzle-calculator_internal_httpserver.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "internal_calculator.CalculateRequest": {
            "type": "object",
            "properties": {
                "a": {
                    "type": "number"
                },
                "b": {
                    "type": "number"
                }
            }
        },
        "internal_calculator.CalculateResponse": {
            "type": "object",
            "properties": {
                "a": {
                    "type": "number"
                },
                "b": {
                    "type": "number"
                },
                "operation": {
                    "$ref": "#/definitions/internal_calculator.Operation"
                },
                "result": {
                    "type": "number"
                }
            }
        },
        "internal_calculator.Operation": {
            "type": "string",
            "enum": [
                "add",
                "subtract",
                "multiply",
                "divide",
                "power",
                "sqrt",
                "percentage"
            ],
            "x-enum-varnames": [
                "OpAdd",
                "OpSubtract",
                "OpMultiply",
                "OpDivide",
                "OpPower",
                "OpSqrt",
                "OpPercentage"
            ]
        }
    }
}